fix: introduce new hook for bulk action
