+++ src/pages/Search/SearchPageWide.tsx
@@ -1,45 +1,37 @@
 import React, {useMemo} from 'react';
+import type {ReactNode} from 'react';
 import type {NativeScrollEvent, NativeSyntheticEvent} from 'react-native';
 import {View} from 'react-native';
 import type {OnyxEntry} from 'react-native-onyx';
 import FullPageNotFoundView from '@components/BlockingViews/FullPageNotFoundView';
-import type {DropdownOption} from '@components/ButtonWithDropdownMenu/types';
 import DragAndDropConsumer from '@components/DragAndDrop/Consumer';
 import DragAndDropProvider from '@components/DragAndDrop/Provider';
 import DropZoneUI from '@components/DropZone/DropZoneUI';
 import ScreenWrapper from '@components/ScreenWrapper';
 import Search from '@components/Search';
 import SearchPageFooter from '@components/Search/SearchPageFooter';
-import SearchFiltersBar from '@components/Search/SearchPageHeader/SearchFiltersBar';
 import SearchPageHeader from '@components/Search/SearchPageHeader/SearchPageHeader';
-import type {SearchHeaderOptionValue} from '@components/Search/SearchPageHeader/SearchPageHeader';
-import type {BankAccountMenuItem, SearchParams, SearchQueryJSON} from '@components/Search/types';
+import type {SearchParams, SearchQueryJSON} from '@components/Search/types';
 import {useMemoizedLazyExpensifyIcons} from '@hooks/useLazyAsset';
 import useLocalize from '@hooks/useLocalize';
 import useTheme from '@hooks/useTheme';
 import useThemeStyles from '@hooks/useThemeStyles';
 import {buildCannedSearchQuery} from '@libs/SearchQueryUtils';
 import Navigation from '@navigation/Navigation';
+import CONST from '@src/CONST';
 import ROUTES from '@src/ROUTES';
 import type {SearchResults} from '@src/types/onyx';
-import type {PaymentMethodType} from '@src/types/onyx/OriginalMessage';
 
 type SearchPageWideProps = {
     queryJSON?: SearchQueryJSON;
     searchResults: OnyxEntry<SearchResults>;
     searchRequestResponseStatusCode: number | null;
     isMobileSelectionModeEnabled: boolean;
-    headerButtonsOptions: Array<DropdownOption<SearchHeaderOptionValue>>;
     footerData: {
         count: number | undefined;
         total: number | undefined;
         currency: string | undefined;
     };
-    selectedPolicyIDs: Array<string | undefined>;
-    selectedTransactionReportIDs: string[];
-    selectedReportIDs: string[];
-    latestBankItems?: BankAccountMenuItem[];
-    onBulkPaySelected: (paymentMethod?: PaymentMethodType) => void;
     handleSearchAction: (value: SearchParams | string) => void;
     onSortPressedCallback: () => void;
     scrollHandler: (event: NativeSyntheticEvent<NativeScrollEvent>) => void;
@@ -47,6 +39,7 @@ type SearchPageWideProps = {
     PDFValidationComponent: React.ReactNode;
     ErrorModal: React.ReactNode;
     shouldShowFooter: boolean;
+    children?: ReactNode;
 };
 
 function SearchPageWide({
@@ -54,13 +47,7 @@ function SearchPageWide({
     searchResults,
     searchRequestResponseStatusCode,
     isMobileSelectionModeEnabled,
-    headerButtonsOptions,
     footerData,
-    selectedPolicyIDs,
-    selectedTransactionReportIDs,
-    selectedReportIDs,
-    latestBankItems,
-    onBulkPaySelected,
     handleSearchAction,
     onSortPressedCallback,
     scrollHandler,
@@ -68,6 +55,7 @@ function SearchPageWide({
     PDFValidationComponent,
     ErrorModal,
     shouldShowFooter,
+    children,
 }: SearchPageWideProps) {
     const styles = useThemeStyles();
     const theme = useTheme();
@@ -83,6 +71,9 @@ function SearchPageWide({
 
     const expensifyIcons = useMemoizedLazyExpensifyIcons(['SmartScan']);
     const handleOnBackButtonPress = () => Navigation.goBack(ROUTES.SEARCH_ROOT.getRoute({query: buildCannedSearchQuery()}));
+
+    // Empty array passed as headerButtonsOptions since SearchBulkActionsButton handles bulk actions
+    const emptyOptions = CONST.EMPTY_ARRAY as [];
 
     return (
         <View style={styles.searchSplitContainer}>
@@ -103,19 +94,11 @@ function SearchPageWide({
                             {PDFValidationComponent}
                             <SearchPageHeader
                                 queryJSON={queryJSON}
-                                headerButtonsOptions={headerButtonsOptions}
+                                headerButtonsOptions={emptyOptions}
                                 handleSearch={handleSearchAction}
                                 isMobileSelectionModeEnabled={isMobileSelectionModeEnabled}
                             />
-                            <SearchFiltersBar
-                                queryJSON={queryJSON}
-                                headerButtonsOptions={headerButtonsOptions}
-                                isMobileSelectionModeEnabled={isMobileSelectionModeEnabled}
-                                currentSelectedPolicyID={selectedPolicyIDs?.at(0)}
-                                currentSelectedReportID={selectedTransactionReportIDs?.at(0) ?? selectedReportIDs?.at(0)}
-                                confirmPayment={onBulkPaySelected}
-                                latestBankItems={latestBankItems}
-                            />
+                            {children}
                             <Search
                                 key={queryJSON.hash}
                                 queryJSON={queryJSON}
